pyproject.toml
[tool.poetry]
name = "hello-world"
version = "0.1.0"
description = "create a single hello world app"
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
pyjwt = "^2.5.0"
bcrypt = "^4.0.0"
fastapi = "*"
prisma = "*"
pydantic = "*"
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
